+++ release-notes.md
@@ -2,8 +2,6 @@ Pending changes in the mainline (upcoming 1.7.0)
 ===============================
 
 Changes:
-  - Disable some UI components on ReadOnly systems.
-  - The study list header is now sticking on top of the screen.
   - When Orthanc DB supports "ExtendedFind" (SQLite in 1.12.5+ and PosgreSQL 7.0+):
     - new features in the local studies list:
       - Allow sorting by columns
@@ -15,6 +13,9 @@ Changes:
       is still used when performing remote DicomWEB queries.
     - New configuration "PageLoadSize" that defines the number of items that are loaded when scrolling the study or instance list.
     - The "StudyListContentIfNoSearch" configuration is now taken into account and always considered as "most-recents".
+  - Disable some UI components on ReadOnly systems.
+  - The study list header is now sticking on top of the screen.
+  - Quick report icon: now display the SeriesDate - SeriesDescription in a tooltip.
 
 Fixes:
   - When modifying studies, dates selected from the DatePicker were not always taken into account.